Add krashna.nl site to showcase
We've recently transitioned from i18n to gatsby-plugin-intl for internalization on our site, and I'd like to add this site to your showcase! We do YAML-based translation, which we merge into one big JSON file for gatsby-plugin-intl to use.

The site I'm PRing is [krashna.nl](https://www.krashna.nl), and is [open-source](https://github.com/krashnamusika/krashna-site). We built and maintain this site for the student orchestra and choir association of Delft (The Netherlands), called Krashna Musika.

Thanks for the great plugin, it's been a timesaver!